Logging and config changes. Improved startup time
