--- config.json
@@ -1,18 +1,0 @@
-{
-	"name": "TestCNC",
-	"TLS": {
-		"privateKey": "testkey.pem",
-		"cert": "testcert.pem"
-	},
-	"Signaller": {
-		"SockAddr": [
-			"localhost:4132",
-			"localhost:4133"
-		],
-		"MOTD": "Authorized use only. Any unauthorised use is strictly prohibited by international law and treaties."
-	},
-	"Web": {
-		"Domain": "localhost",
-		"Listener": "0.0.0.0:4132"
-	}
-}
