+++ src/github.com/twitchyliquid64/CNC/data/main.go
@@ -1,10 +1,6 @@
 package data
 
 import (
-  "github.com/twitchyliquid64/CNC/data/stmdata"
-  "github.com/twitchyliquid64/CNC/data/session"
-  "github.com/twitchyliquid64/CNC/data/entity"
-  "github.com/twitchyliquid64/CNC/data/plugin"
   "github.com/twitchyliquid64/CNC/data/user"
   "github.com/twitchyliquid64/CNC/logging"
   "github.com/twitchyliquid64/CNC/config"
@@ -38,7 +34,7 @@ func Initialise() {
     logging.Error("data", "Error: ", err)
   }
 
-  checkStructures()
+  checkStructures(DB)
 
   //make sure that objects in the config BaseObjects are
   //existing, creating them if nessesary.
@@ -61,51 +57,4 @@ func Initialise() {
   }
 
   logging.Info("data", "Initialisation finished.")
-}
-
-func autoMigrateTables() {
-  logging.Info("data", "Checking structure: Users")
-  DB.AutoMigrate(&user.User{})
-  logging.Info("data", "Checking structure: Permissions")
-  DB.AutoMigrate(&user.Permission{})
-  user.Permission{}.Init(DB)
-  logging.Info("data", "Checking structure: Emails")
-  DB.AutoMigrate(&user.Email{})
-  logging.Info("data", "Checking structure: Addresses")
-  DB.AutoMigrate(&user.Address{})
-  logging.Info("data", "Checking structure: AuthenticationMethods")
-  DB.AutoMigrate(&user.AuthenticationMethod{})
-  logging.Info("data", "Checking structure: Sessions")
-  DB.AutoMigrate(&session.Session{})
-
-  logging.Info("data", "Checking structure: Entity")
-  DB.AutoMigrate(&entity.Entity{})
-  logging.Info("data", "Checking structure: EntityPivot")
-  DB.AutoMigrate(&entity.EntityPivot{})
-  logging.Info("data", "Checking structure: EntityLocationRecord")
-  DB.AutoMigrate(&entity.EntityLocationRecord{})
-  logging.Info("data", "Checking structure: EntityStatusRecord")
-  DB.AutoMigrate(&entity.EntityStatusRecord{})
-  logging.Info("data", "Checking structure: EntityLogRecord")
-  DB.AutoMigrate(&entity.EntityLogRecord{})
-  logging.Info("data", "Checking structure: EntityEvent")
-  DB.AutoMigrate(&entity.EntityEvent{})
-
-  logging.Info("data", "Checking structure: Plugin")
-  DB.AutoMigrate(&plugin.Plugin{})
-  logging.Info("data", "Checking structure: Resource")
-  DB.AutoMigrate(&plugin.Resource{})
-  logging.Info("data", "Checking structure: Stmdata")
-  DB.AutoMigrate(&stmdata.Stmdata{})
-}
-
-//called during initialisation. Should make sure the schema is intact and up to date.
-func checkStructures() {
-  logging.Info("Auto migrating tables")
-  autoMigrateTables()
-
-  logging.Info("Migrating DB")
-  upgradeDb(DB)
-
-  logging.Info("Migrations finished")
 }
